requirements.txt
pandas
numpy
psycopg2-binary
torch
scikit-learn
matplotlib 